So many
Its been so long since the last commit, all i can say is that im mostly working on the frontend.

Im working on the app part of this where we make schedules and such. Currently have the sliders available, and need to add ways to add and subtract users. Submit schedule button too,.