Can remove sliders
So you can remove sliders by clicking the name.

Trying to start adding users by clicking the plus.

Will later make it so that users removed from the [active] list get put in the [inactive] list in state. The inactives are in the nav and the actives are in the sliders

Also trying to organize css