Sending Data to Python
So when you click save changes it sends a post request with the slider info as well as user id. The request can now be sent, but now i have to wonder how it should be stored.

I could have week objects and just have days in them and each day can have a bunch of 'slots' that have a user_id, and the start and end times.

Also this could just be stored inside each user? IDK