Refactoring and More
So sliders are not a separate array in state, they are a prop of user now  and stored as value. For instance, users[0].value   >> ['08:00', '16:00'].

Next im trying to have the sliders default read from the user values.

Im still having the problem where the wrong sliders get deleted, but im trying to solve it seemlessly and not make another function that resets everything when a slider is deleted but it might come to that.