Fixed and added
Fixed all the slider issues thanks to stackoverflow.

Starting to build some of the backend infrustructure involved with organizing the data submitted from react.
Workblocks, Day added. I may need a week object or something though.
When user submits, it sends the date to Python, as well as each user and their slider data. This is stored as workblocks which can later be retrieved.