More Flask Stuff
So Ive added links to the day circles. They send their date to the route, the route searches for a day that matches, then posts a req to flask. Flask searches db for a day that matches, and if it cant find it makes a new day.

Im just cleaning up some of the errors thought right now.