Welp
So it looks like since React is for making webapps, i cannot set the state of components during anything really exept for events. The api call or event can happen on a button press,  or on startup of application, or on a timer. Not when a url is submitted tho.