Login Logic
So now a login tests for username and if it cant find it it shows and alert message.

If it does find the user it sends the req to flask and if the password is wrong flask returns an unauthenticated user.

If react sees that the user isn't authenticated, it adds a password error message.