Progress
If user trys to login with their name, forward to register page (im going to implement this)

Need feedback when user is added

Now must be a manager to access the add user, and for the button to show up in the drawer

I could either steer into this and try to fix all that or go and work on other components.

Next would be the profile page.