Added Availability Form
Fixed a lot about the User page. Made it so that if you're a manager you can edit but if not you have to be that crew member to edit info.

In the availability form, i made the switch change between inactive slider and active slider. Just need to set up values.