Trouble With Get Req
So you can now save your availability and the db will reflect this. Question is moreso how do you convert the time in the db back to value so the sliders can use it. I have created a function for this but i have some errors. To remove errors just comment out the function and leave return [0,50]