Created loadWorkBlocksIn
Still on that thing from yesterday.
I need to either convert the Day object i get from Python to a useable Day object in JS either by just using the same types of names (first_name and firstName) or just convert when i make the request.

I think i have to do the latter.

I started to make a function that i will call in state when loading the component that will unload a bunch of wbs. Its called loadWorkBlockIn.

So go to that and to ShiftView for what i said above.