Schedulepaper fixed
Fixed Schedulepaper and made it responsive to either dashboard or whatever it needs to be on.

Working on the icon btns. Thinking of using an svg line and animating it up and down.

Next is that then the editor. Once the editor is done, then saving, then we're polishing. Editor will likely take a while though.