Time Crap
So as of recent ive been using momentjs. Problem is converting back and forth from moment to whatever is really annoying soooo what we'll do is we will overall store the time data in state or props as a string (isoformat) and whenever we need to operate on it we'll pass it to a function and convert it to a moment. This way we never have to guess as to what kind of data the argument 'time' may be