Changes
So i have to move the state functionality from Timeslot to MainContent in order for me to submit it...

I need to convert each workblock's time to pix when they get requested, which requires knowlege of the shiftfilter. So i  need to move shiftfilte functionality from vizualizer to MainContent.

That's a lot but it is neccessary. I think Redux would help with this however... im not learning it